緯度経度の単位を変更

+++ create_weathers.sql
@@ -2,8 +2,8 @@ create table weathers (
   id int primary key, 
   pref varchar(32),
   area varchar(32),
-  latitude double,
-  longitude double,
+  latitude decimal,
+  longitude decimal,
   url varchar(64),
   xpath varchar(64)
 )
